Revert "update two players's files so it follows LSP"
