+++ src/main2/use_case/two_player/PlayInteractorAdapter.java
@@ -2,7 +2,6 @@ package main2.use_case.two_player;
 
 import main2.use_case.play.PlayInputBoundary;
 import main2.use_case.play.PlayInputData;
-import main2.use_case.play.OnePlayerInteractor;
 import main2.use_case.play.PlayInteractor;
 import main2.use_case.settings.SettingsDTO;
 import main2.use_case.settings.SettingsInteractor;
@@ -11,12 +10,12 @@ import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 
 public class PlayInteractorAdapter implements PlayInputBoundary, PropertyChangeListener {
-    private final OnePlayerInteractor playInteractor;
+    private final PlayInteractor playInteractor;
     private final TwoPlayerInteractor twoPlayerInteractor;
+
     private PlayInteractor activeInteractor;
 
-
-    public PlayInteractorAdapter(OnePlayerInteractor playInteractor,
+    public PlayInteractorAdapter(PlayInteractor playInteractor,
                                  TwoPlayerInteractor twoPlayerInteractor,
                                  SettingsInteractor settingsInteractor) {
         this.playInteractor = playInteractor;
